Proof of concept VM agnostic API
This API covers an example of what the VM agnostic API for native
modules might look like.  In particular it currently covers the first
two node native module examples, hello world and function arguments.

It is not intended to be the final design.  Indeed it is far from it,
skipping over many details like error handling and providing only opaque
types meaning all operations must go across the DLL boundary which is
expected to be a performance issue.