Add get holder and object wrap handle to NAPI
Added napi_get_cb_holder().  The holder is a V8 API concept, but it
seems to be particularly useful, and is worth considering making a
permanent part of the API that other VMs would have to emulate or
implement.

Renamed napi_get_cb_object to napi_get_cb_this to be consistent with
This and Holder concepts.

Added a napi_persistent out parameter to napi_wrap().  Turns out
ObjectWrap maintains a persistent handle to the JavaScript object that
is wrapping the native object.  This was made easy by inheriting your
wrapped object from ObjectWrap, but with the C API we cannot do that.
So instead a manual out parameter for a persistent is what we can do.
It is currently optional, ignored if NULL.