Introduce handle scope APIs and helper C++ classes
New APIs for handling HandleScope and EscapleHandleScope model.  Other
VMs may not need these and would implement them as no-ops.  E.g. VMs
with conservative GCs like Chakra.

This also introduces a C++ namespace, Napi, which is to include the C++
helper layer.  This namespace must all be inline in the header file and
not use symbol exports to ensure we have a clean C API with stable ABI.