Add a very rough API for abstracting v8::TryCatch
Adds napi_try_catch which takes two callbacks, one for the try body, and
one for the catch body.  It then returns bool for whether it had caught
(and terminated propagation of) an exception or not.

This api has not been given enough thought.  The callback approach is
necessary with the current v8::TryCatch implementation because it
privately acquires the current stack position where it is constructed,
which would be incorrect if taken from an API designed to returne a
token handle to a trycatch descriptor and require explicit release with
a later api call.  This design would capture the stack position of that
first api call, which could then overlap with other calls to javascript
code that may also have try/catch handlers.

The callback approach is nice but adds significant overhead to the
caller's code since it must be wrapped up into one or two callbacks. It
is also unclear without further research what comprehensive
functionality is needed by this api.

It also requires a manual callback closure via the void* data parameter
which further makes the caller's code verbose and cumbersome.  This is
probably unavoidable via the callback approach since we want to remain a
C API.