Rename napi_throw_error to napi_throw
napi_throw_error is misleading since you can legitimately throw any
napi_value (any javascript value).  So renamed to napi_throw to clarify.
This also leaves room for a napi_throw_error which would be a helper
that creates a new error from a string and throws it for you, to
collapse the otherwise three calls that would be needed into one.

Also renamed a bunch of parameters named napi_env to just env for
clarity.