Add napi_create_constructor_for_wrap_with_methods
This api takes a name and napi_method_descriptors and creates a
constructor function with that name and those methods as its prototype
methods.

This allows the V8 implementation to take advantage of
v8::FunctionTemplate for a significant performance boost.  It also
improves the legibility of the client code using this API.