Initial merge of node-chakracore/api-opaque-prototype
This work supports most of the napi set of API with Jsrt-backed
implementations instead of relying on the chakrashim layer. There are
currently two issues blocking the leveldown module unit test suite.
First is an issue with how weak references are exposed from Jsrt and
second is an issue in the prototype assigned to buffer objects created
via the napi layer.

The base node-chakracore commit of this checkin is nodejs/node-chakracore/xplat commit 9cb6342ef9672129a982d9aea8783804fe96ea6e