napi: Remove napi_propertyname
This change removes ```napi_propertyname``` and replaces their use with ```napi_value```. On v8 ```napi_propertyname``` is just a ```v8::Local<v8::String>``` anyway. It doesn't seem to be worthwhile from the types of code patterns we've seen in modules to keep the separate type for property keys.

Also introduces property get/set/has for properties named directly by a ```const char*``` which seems to be the largest use of property access on objects. Naming the property key via a utf8 string is keeping more in-line with the other napi method calls for dealing with names of things like ```napi_property_descriptor``` or the ```const char*``` we use to name classes (via ```napi_define_class```) or functions (via ```napi_create_function```).

Adds a few simple tests of these APIs.